+++ bzt/modules/shellexec.py
@@ -22,7 +22,7 @@ from bzt.six import iteritems
 from bzt.utils import ensure_is_dict
 from bzt.utils import shutdown_process, BetterDict, is_windows
 
-ARTIFACTS_DIR_EVVAR = "TAURUS_ARTIFACTS_DIR"
+ARTIFACTS_DIR_ENVVAR = "TAURUS_ARTIFACTS_DIR"
 
 
 class ShellExecutor(Service):
@@ -59,7 +59,7 @@ class ShellExecutor(Service):
                 env.merge({"PYTHONPATH": working_dir})
                 if os.getenv("PYTHONPATH"):
                     env['PYTHONPATH'] = os.getenv("PYTHONPATH") + os.pathsep + env['PYTHONPATH']
-                env[ARTIFACTS_DIR_EVVAR] = self.engine.artifacts_dir
+                env[ARTIFACTS_DIR_ENVVAR] = self.engine.artifacts_dir
 
                 for name, value in iteritems(env):
                     env[str(name)] = str(value)