+++ site/dat/docs/Changelog.md
@@ -2,7 +2,8 @@
 
 ## 1.6.3 (next)
  - fix percentile value handling in passfail criteria
- - fix problem with sending included configs into cloud 
+ - fix problem with sending included configs into cloud
+ - fix cumulative fail criteria processing order
 
 ## 1.6.2 <sup>8 jun 2016</sup>
  - fix passfail-related regression crash