+++ site/dat/docs/Changelog.md
@@ -4,6 +4,8 @@
  - fix percentile value handling in passfail criteria
  - fix problem with sending included configs into cloud 
  - add setting up of CWD on server side
+ - fix problem with sending included configs into cloud
+ - fix cumulative fail criteria processing order
 
 ## 1.6.2 <sup>8 jun 2016</sup>
  - fix passfail-related regression crash